Fix "wrong type" warnings

+++ src/pixels.c
@@ -699,15 +699,15 @@ PHP_FUNCTION(SDL_SetPaletteColors)
 	if (ncolors <= 0) {
 		ncolors = count;
 	} else if (ncolors > count) {
-		php_error_docref(NULL, E_NOTICE, "Invalid color number, %ld will be used instead of %ld", count, ncolors);
+		php_error_docref(NULL, E_NOTICE, "Invalid color number, %ld will be used instead of %ld", count, (long) ncolors);
 		ncolors = count;
 	}
 	if (first >= intern->palette->ncolors) {
-		php_error_docref(NULL, E_WARNING, "Invalid first color index, %ld >= %d", first, intern->palette->ncolors);
+		php_error_docref(NULL, E_WARNING, "Invalid first color index, %ld >= %d", (long) first, intern->palette->ncolors);
 		RETURN_LONG(-1);
 	}
 	if (first+ncolors > intern->palette->ncolors) {
-		php_error_docref(NULL, E_WARNING, "Invalid color number, %ld > %d", first+ncolors, intern->palette->ncolors);
+		php_error_docref(NULL, E_WARNING, "Invalid color number, %ld > %d", (long) (first+ncolors), intern->palette->ncolors);
 		RETURN_LONG(-1);
 	}
 	colors = emalloc(sizeof(SDL_Color) * ncolors);
@@ -1137,7 +1137,7 @@ PHP_METHOD(SDL_Pixels, GetByte)
 	intern = PHP_SDL_PIXELS_P(z_pixels);
 
 	if (x < 0 || x >= intern->pixels.pitch || y < 0 || y >= intern->pixels.h) {
-		php_error_docref(NULL, E_NOTICE, "Invalid position (%ld,%ld) in SDL_Pixels (%d,%d)", x, y, intern->pixels.pitch, intern->pixels.h);
+		php_error_docref(NULL, E_NOTICE, "Invalid position (%ld,%ld) in SDL_Pixels (%d,%d)", (long) x, (long) y, intern->pixels.pitch, intern->pixels.h);
 		RETURN_FALSE;
 	}
 	RETVAL_LONG(intern->pixels.pixels[y*intern->pixels.pitch+x]);
@@ -1164,7 +1164,7 @@ PHP_METHOD(SDL_Pixels, SetByte)
 	intern = PHP_SDL_PIXELS_P(z_pixels);
 
 	if (x < 0 || x >= intern->pixels.pitch || y < 0 || y >= intern->pixels.h) {
-		php_error_docref(NULL, E_NOTICE, "Invalid position (%ld,%ld) in SDL_Pixels (%d,%d)", x, y, intern->pixels.pitch, intern->pixels.h);
+		php_error_docref(NULL, E_NOTICE, "Invalid position (%ld,%ld) in SDL_Pixels (%d,%d)", (long) x, (long) y, intern->pixels.pitch, intern->pixels.h);
 		RETURN_FALSE;
 	}
 	RETVAL_LONG(intern->pixels.pixels[y*intern->pixels.pitch+x]);
